requirements.txt
gradio
numpy
pillow
scikit-image
tensorflow
keras